Start puzzle hero

+++ api/src/modules/redis/redis.service.ts
@@ -47,7 +47,9 @@ export class RedisService {
                 }
             });
             stream.on('end', async () => {
-                await this.client.del(...keys);
+                if (keys && keys.length) {
+                    await this.client.del(...keys);
+                }
                 resolve();
             });
         });
